Remove unused types/files
