--- packages/nimble-blazor/NI.Nimble.Blazor.Components/ColumnDefinition.cs
@@ -1,21 +1,0 @@
-﻿using System.Linq.Expressions;
-
-namespace NI.Nimble.Components;
-
-public class ColumnDefinition<TItem>
-{
-    public string ColumnDataKey { get; set; }
-    public string Title { get; set; }
-
-    public Func<TItem, object>? FieldSelector { get; set; }
-
-    public Expression<Func<TItem, object>>? FieldSelectorExpression { get; set; }
-
-    public ColumnDefinition(string fieldName, Expression<Func<TItem, object>> fieldSelectorExpression)
-    {
-        ColumnDataKey = fieldName;
-        Title = fieldName;
-        FieldSelectorExpression = fieldSelectorExpression;
-        FieldSelector = fieldSelectorExpression.Compile();
-    }
-}
