--- packages/nimble-blazor/NI.Nimble.Blazor.Components/DataGridCellType.cs
@@ -1,16 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum DataGridCellType
-{
-    Default,
-    ColumnHeader,
-    RowHeader
-}
-
-internal static class CellTypeExtensions
-{
-    private static readonly Dictionary<DataGridCellType, string> _cellTypeValues =
-        Enum.GetValues<DataGridCellType>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this DataGridCellType? value) => value == null ? null : _cellTypeValues[value.Value];
-}
