--- packages/nimble-blazor/NI.Nimble.Blazor.Components/GenerateHeaderOptions.cs
@@ -1,16 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum GenerateHeaderOptions
-{
-    None,
-    Default,
-    Sticky
-}
-
-internal static class GenerateHeaderExtensions
-{
-    private static readonly Dictionary<GenerateHeaderOptions, string> _generateHeaderValues =
-        Enum.GetValues<GenerateHeaderOptions>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this GenerateHeaderOptions? value) => value == null ? null : _generateHeaderValues[value.Value];
-}
