--- packages/nimble-blazor/NI.Nimble.Blazor.Components/HorizontalPosition.cs
@@ -1,18 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum HorizontalPosition
-{
-    Unset,
-    Start,
-    End,
-    Left,
-    Right
-}
-
-internal static class HorizontalDefaultPositionExtensions
-{
-    private static readonly Dictionary<HorizontalPosition, string> _positionValues =
-        Enum.GetValues<HorizontalPosition>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this HorizontalPosition? value) => value == null ? null : _positionValues[value.Value];
-}
