--- packages/nimble-blazor/NI.Nimble.Blazor.Components/LocalizationDirection.cs
@@ -1,16 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum LocalizationDirection
-
-{
-    ltr,
-    rtl
-}
-
-internal static class LocalizationDirectionExtensions
-{
-    private static readonly Dictionary<LocalizationDirection, string> _directionValues =
-        Enum.GetValues<LocalizationDirection>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this LocalizationDirection? value) => value == null ? null : _directionValues[value.Value];
-}
