--- packages/nimble-blazor/NI.Nimble.Blazor.Components/Orientation.cs
@@ -1,15 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum Orientation
-{
-    Horizontal,
-    Vertical
-}
-
-internal static class OrientationExtensions
-{
-    private static readonly Dictionary<Orientation, string> _orientationValues =
-        Enum.GetValues<Orientation>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this Orientation? value) => value == null ? null : _orientationValues[value.Value];
-}
