--- packages/nimble-blazor/NI.Nimble.Blazor.Components/Resize.cs
@@ -1,16 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum Resize
-{
-    Horizontal,
-    Vertical,
-    Both
-}
-
-internal static class ResizeExtensions
-{
-    private static readonly Dictionary<Resize, string> _resizeValues =
-        Enum.GetValues<Resize>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this Resize? value) => value == null ? null : _resizeValues[value.Value];
-}
