--- packages/nimble-blazor/NI.Nimble.Blazor.Components/Shape.cs
@@ -1,15 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum Shape
-{
-    Rect,
-    Circle
-}
-
-internal static class ShapeExtensions
-{
-    private static readonly Dictionary<Shape, string> _orientationValues =
-        Enum.GetValues<Shape>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this Shape? value) => value == null ? null : _orientationValues[value.Value];
-}
