--- packages/nimble-blazor/NI.Nimble.Blazor.Components/UpdateMode.cs
@@ -1,15 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum UpdateMode
-{
-    Anchor,
-    Auto
-}
-
-internal static class UpdateModeExtensions
-{
-    private static readonly Dictionary<UpdateMode, string> _updateModeValues =
-        Enum.GetValues<UpdateMode>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this UpdateMode? value) => value == null ? null : _updateModeValues[value.Value];
-}
