--- packages/nimble-blazor/NI.Nimble.Blazor.Components/VerticalPosition.cs
@@ -1,16 +1,0 @@
-﻿namespace NI.Nimble.Components;
-
-public enum VerticalPosition
-{
-    Unset,
-    Top,
-    Bottom
-}
-
-internal static class VerticalDefaultPositionExtensions
-{
-    private static readonly Dictionary<VerticalPosition, string> _positionValues =
-        Enum.GetValues<VerticalPosition>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
-
-    public static string? ToAttributeValue(this VerticalPosition? value) => value == null ? null : _positionValues[value.Value];
-}
